chore: rename sound project
