+++ SampleProjects/Guides/Audio/stylecop.json
